move components to where they are used
