+++ x-pack/platform/plugins/shared/onechat/public/application/components/conversations/conversation_input/agent_select_dropdown.tsx
@@ -22,9 +22,9 @@ import {
   useEuiTheme,
 } from '@elastic/eui';
 import type { AgentDefinition } from '@kbn/onechat-common';
-import { appPaths } from '../../utils/app_paths';
-import { useNavigation } from '../../hooks/use_navigation';
-import { labels } from '../../utils/i18n';
+import { appPaths } from '../../../utils/app_paths';
+import { useNavigation } from '../../../hooks/use_navigation';
+import { labels } from '../../../utils/i18n';
 
 const agentSelectId = 'agentBuilderAgentSelect';
