+++ x-pack/platform/plugins/shared/onechat/public/application/components/conversations/conversation_input/conversation_input_actions.tsx
@@ -7,7 +7,7 @@
 
 import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import React from 'react';
-import { ConversationAgentSelector } from '../conversation_agent_selector';
+import { ConversationAgentSelector } from './conversation_agent_selector';
 import { ConversationActionButton } from './conversation_action_button';
 import { useSendMessage } from '../../../context/send_message/send_message_context';
 import { ConnectorSelector } from '../../connector_selector';
